Moving pieces (unvalidated)
Pieces can now be moved in the game screen by clicking on a piece and then its destination. There is currently no checking that a move is legal (although moves from the board to storage and storage to storage do not work as I deliberately did not implement the checks to allow that).